+++ CashApp/CashApp/Services/IRestService.cs
@@ -1,21 +1,22 @@
 ﻿using CashApp.Models;
 using System.Collections.Generic;
+using System.Threading;
 using System.Threading.Tasks;
 
 namespace CashApp.Services
 {
     public interface IRestService
     {
-        Task<List<TransactionWithPeriod>> GetPeriodData();
+        Task<List<TransactionWithPeriod>> GetPeriodData(CancellationTokenSource cts);
 
-        Task<List<Transaction>> GetPeriodData(int year, int month);
+        Task<List<Transaction>> GetPeriodData(int year, int month, CancellationTokenSource cts);
 
-        Task<List<Transaction>> GetAllData();
+        Task<List<Transaction>> GetAllData(CancellationTokenSource cts);
 
-        Task<Transaction> GetData(int id);
+        Task<Transaction> GetData(int id, CancellationTokenSource cts);
 
-        Task<bool> SaveItem(Transaction item);
+        Task<bool> SaveItem(Transaction item, CancellationTokenSource cts);
 
-        Task DeleteItem(int id);
+        Task DeleteItem(int id, CancellationTokenSource cts);
     }
 }